requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
kubernetes>=30.1.0
ollama>=0.3.3
openai>=1.14.0
anthropic>=0.21.3
sqlalchemy[asyncio]>=2.0.28
alembic>=1.13.1
asyncpg>=0.29.0
celery>=5.3.6
redis>=5.0.3
prometheus_client>=0.20.0
structlog>=24.2.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
pyyaml>=6.0.1
websockets>=12.0
pytest>=8.3.3
pytest-asyncio>=0.23.5
pytest-cov>=5.0.0
PyJWT>=2.8.0